+++ backup/strategy_mysql.go
@@ -44,8 +44,10 @@ func loadMysqlMaster(cli *client.Client, serviceID string, event *csevent.Projec
 
 	if err != nil {
 		backupLogger().Warn("Failed to load MySQL Master", "error", err.Error(), "function", "loadMysqlMaster", "action", "FindRunningContainer", "serviceID", serviceID)
-		go event.PostEventUpdate("agent-6e0d0cbb268ac0b1", "Container appears to be offline. Unable to perform"+" "+event.EventLog.Locale)
-		event.EventLog.Status = "cancelled"
+		if event != nil {
+			go event.PostEventUpdate("agent-6e0d0cbb268ac0b1", "Container appears to be offline. Unable to perform"+" "+event.EventLog.Locale)
+			event.EventLog.Status = "cancelled"
+		}
 		return &MysqlInstance{}, err
 	}
 
@@ -118,8 +120,10 @@ func loadMysqlMaster(cli *client.Client, serviceID string, event *csevent.Projec
 
 	if versionStage == "" {
 		backupLogger().Warn("Failed to identify MySQL Version", "error", "version string is blank", "function", "loadMysqlMaster", "serviceID", serviceID)
-		go event.PostEventUpdate("agent-f422717152297b23", "Unable to load MySQL Version, halting job. "+" "+event.EventLog.Locale)
-		event.EventLog.Status = "failed"
+		if event != nil {
+			go event.PostEventUpdate("agent-f422717152297b23", "Unable to load MySQL Version, halting job. "+" "+event.EventLog.Locale)
+			event.EventLog.Status = "failed"
+		}
 		return &instance, errors.New("missing version string")
 	}
 
@@ -137,8 +141,10 @@ func loadMysqlMaster(cli *client.Client, serviceID string, event *csevent.Projec
 	v, vErr := semver.NewVersion(versionStage)
 	if vErr != nil {
 		backupLogger().Warn("Failed to identify MySQL Version", "error", vErr.Error(), "function", "loadMysqlMaster", "serviceID", serviceID)
-		go event.PostEventUpdate("agent-114697045cd756da", "Unable to load MySQL Version, halting job.\n\n"+vErr.Error())
-		event.EventLog.Status = "failed"
+		if event != nil {
+			go event.PostEventUpdate("agent-114697045cd756da", "Unable to load MySQL Version, halting job.\n\n"+vErr.Error())
+			event.EventLog.Status = "failed"
+		}
 		return &instance, vErr
 	}